Shifting Navigation staff in DashboardActivity
