+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/core/MifosBaseActivity.java
@@ -57,16 +57,12 @@ import rx.schedulers.Schedulers;
 /**
  * @author fomenkoo
  */
-public class MifosBaseActivity extends AppCompatActivity implements BaseActivityCallback,
-        NavigationView.OnNavigationItemSelectedListener, CompoundButton.OnCheckedChangeListener {
+public class MifosBaseActivity extends AppCompatActivity implements BaseActivityCallback {
 
     protected Toolbar toolbar;
     private ActivityComponent mActivityComponent;
     private ProgressDialog progress;
-    private NavigationView mNavigationView;
-    private DrawerLayout mDrawerLayout;
-    private DataManager mDataManager;
-    private SwitchCompat offline_toggle;
+
 
     @Override
     public void setContentView(int layoutResID) {
@@ -162,195 +158,5 @@ public class MifosBaseActivity extends AppCompatActivity implements BaseActivity
             }
             transaction.commit();
         }
-    }
-
-    @Override
-    public void onBackPressed() {
-        // check if the nav mDrawer is open
-        if (mDrawerLayout != null && mDrawerLayout.isDrawerOpen(GravityCompat.START)) {
-            mDrawerLayout.closeDrawer(Gravity.LEFT);
-        } else {
-            super.onBackPressed();
-        }
-
-    }
-
-    @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-
-        // ignore the current selected item
-        /*if (item.isChecked()) {
-            mDrawerLayout.closeDrawer(Gravity.LEFT);
-            return false;
-        }*/
-
-        // select which activity to open
-        final Intent intent = new Intent();
-        switch (item.getItemId()) {
-            case R.id.item_dashboard:
-                replaceFragment(new ClientSearchFragment(), false, R.id.container);
-                break;
-            case R.id.item_clients:
-                replaceFragment(new ClientListFragment(), false, R.id.container);
-                break;
-            case R.id.item_groups:
-                replaceFragment(new GroupsListFragment(), false, R.id.container);
-                break;
-            case R.id.item_centers:
-                intent.setClass(getApplicationContext(), CentersActivity.class);
-                startNavigationClickActivity(intent);
-                break;
-            case R.id.item_survey:
-                intent.setClass(getApplicationContext(), SurveyActivity.class);
-                startNavigationClickActivity(intent);
-                break;
-            case R.id.item_path_tracker:
-                intent.setClass(getApplicationContext(), PathTrackingActivity.class);
-                startNavigationClickActivity(intent);
-                break;
-            case R.id.item_offline:
-                intent.setClass(getApplicationContext(), OfflineCenterInputActivity.class);
-                startNavigationClickActivity(intent);
-                break;
-
-        }
-
-        // close the drawer
-        mDrawerLayout.closeDrawer(Gravity.LEFT);
-        mNavigationView.setCheckedItem(R.id.item_dashboard);
-        return true;
-    }
-
-    /**
-     * sets up the navigation mDrawer in the activity
-     */
-    protected void setupNavigationBar() {
-
-        // setup navigation view
-        mNavigationView = (NavigationView) findViewById(R.id.navigation_view);
-        mNavigationView.setNavigationItemSelectedListener(this);
-
-        // setup drawer layout and sync to toolbar
-        mDrawerLayout = (DrawerLayout) findViewById(R.id.drawer);
-        ActionBarDrawerToggle actionBarDrawerToggle = new ActionBarDrawerToggle(this,
-                mDrawerLayout, toolbar, R.string.open_drawer, R.string.close_drawer) {
-
-            @Override
-            public void onDrawerClosed(View drawerView) {
-                super.onDrawerClosed(drawerView);
-            }
-
-            @Override
-            public void onDrawerOpened(View drawerView) {
-                super.onDrawerOpened(drawerView);
-            }
-        };
-        mDrawerLayout.setDrawerListener(actionBarDrawerToggle);
-        actionBarDrawerToggle.syncState();
-
-        // make an API call to fetch logged in client's details
-        loadClientDetails();
-        setupOfflineSwitchCompat();
-    }
-
-    public void setupOfflineSwitchCompat() {
-        android.support.v7.widget.SwitchCompat offline_toggle
-                = (android.support.v7.widget.SwitchCompat) findViewById(R.id.offline_switch_compat);
-        /*if (PrefManager.getUserStatus() == Constants.USER_OFFLINE)
-            offline_toggle.setChecked(true);*/
-    }
-
-    @Override
-    public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-        if (PrefManager.getUserStatus() == Constants.USER_OFFLINE) {
-            PrefManager.setUserStatus(Constants.USER_ONLINE);
-            //offline_toggle.setChecked(false);
-        } else {
-            PrefManager.setUserStatus(Constants.USER_OFFLINE);
-            //offline_toggle.setChecked(true);
-        }
-    }
-
-    /**
-     * downloads the client name and picture(if exists)
-     * sets the downloaded data to the nav drawer account header
-     */
-    //TODO Setup In MVP
-    private void loadClientDetails() {
-
-        // download client details
-        final int userId = PrefManager.getUserId();
-        BaseApiManager baseApiManager = new BaseApiManager();
-        mDataManager = new DataManager(baseApiManager);
-        Observable<Client> call = mDataManager.getClient(userId);
-        Subscription subscription = call.subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Subscriber<Client>() {
-                    @Override
-                    public void onCompleted() {
-
-                    }
-
-                    @Override
-                    public void onError(Throwable e) {
-
-                    }
-
-                    @Override
-                    public void onNext(Client client) {
-                        // add name to profile
-                        String name = client.getDisplayName();
-                        TextView textViewUsername = (TextView) findViewById(R.id.tv_user_name);
-                        textViewUsername.setText(name);
-
-                        // download image
-                        if (client.isImagePresent()) {
-                            String url = PrefManager.getInstanceUrl()
-                                    + "/"
-                                    + "clients/"
-                                    + userId
-                                    + "/images?maxHeight=120&maxWidth=120";
-                            GlideUrl glideUrl = new GlideUrl(url, new LazyHeaders.Builder()
-                                    .addHeader(MifosInterceptor.HEADER_TENANT, PrefManager
-                                            .getTenant())
-                                    .addHeader(MifosInterceptor.HEADER_AUTH, PrefManager
-                                            .getToken())
-                                    .addHeader("Accept", "application/octet-stream")
-                                    .build());
-
-                            ImageView imageViewUserPicture = (ImageView) findViewById(R.id
-                                    .iv_user_picture);
-                            Glide.with(getApplicationContext())
-                                    .load(glideUrl)
-                                    .asBitmap()
-                                    .error(R.drawable.ic_account_circle)
-                                    .placeholder(R.drawable.ic_account_circle)
-                                    .into(new BitmapImageViewTarget(imageViewUserPicture) {
-                                        @Override
-                                        protected void setResource(Bitmap result) {
-                                            // check a valid bitmap is downloaded
-                                            if (result == null || result.getWidth() == 0)
-                                                return;
-
-                                            // set to image view
-                                            ImageView imageViewUserPicture = (ImageView)
-                                                    findViewById(R.id.iv_user_picture);
-                                            imageViewUserPicture.setImageBitmap(result);
-                                        }
-                                    });
-                        }
-                    }
-                });
-    }
-
-    public void startNavigationClickActivity(final Intent intent) {
-        android.os.Handler handler = new android.os.Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                startActivity(intent);
-
-            }
-        }, 500);
     }
 }
